requirements.txt
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
colorama==0.4.6
distlib==0.3.6
Django==4.1.7
django-bootstrap-icons==0.8.2
django-bootstrap-v5==1.0.11
django-modeltranslation==0.18.8
django-parler==2.3
django-rosetta==0.9.8
filelock==3.9.0
idna==3.4
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
polib==1.1.1
psycopg2-binary==2.9.5
pyproject_api==1.5.0
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
tox==4.4.4
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.18.0
